local_requirements.txt
impacket>=0.12.0
smbprotocol>=1.15.0
rich>=14.0.0
colorama>=0.4.6
yara-python>=4.3.0
termcolor>=2.3.0
prettytable>=3.8.0
python-dotenv>=1.0.0
cryptography>=42.0.0
pyasn1>=0.4.8